move basic model entity GQL types into constants package
