+++ packages/constants/src/index.ts
@@ -4,5 +4,6 @@ export * from './dateFnsFormats';
 export * from './datetime';
 export * from './defaultDates';
 export * from './misc';
+export * from './modelInterfaces';
 export * from './time';
 export * from './ticket';
