+++ workspaces.json
@@ -32,10 +32,6 @@
         "location": "packages/adapters"
     },
     {
-        "name": "@eventespresso/config",
-        "location": "packages/config"
-    },
-    {
         "name": "@eventespresso/constants",
         "location": "packages/constants"
     },
